backward module exports compatibility
