+++ bench/Cargo.toml
@@ -8,8 +8,8 @@ version = "0.1.0"
 crate-type = ["cdylib"]
 
 [dependencies]
-napi = {path = "../napi", features = ["tokio_rt", "serde-json"]}
-napi-derive = {path = "../napi-derive"}
+napi = {path = "../napi", features = ["tokio_rt", "serde-json", "compat-mode"]}
+napi-derive = {path = "../crates/macro", features = ["compat-mode"]}
 serde = "1"
 serde_json = "1"
