+++ crates/backend/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "backend"
+name = "napi-macro-backend"
 version = "0.1.0"
 edition = "2018"
