--- crates/macro-example/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-authors = ["LongYinan <lynweklm@gmail.com>"]
-edition = "2018"
-name = "macro-example"
-version = "0.1.0"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-napi-macro = { path = "../macro" }
-napi = { path = "../../napi", features = ["bindgen", "latin1"] }
-
-[build-dependencies]
-napi-build = { path = "../../build" }
