+++ crates/macro/src/parser/attrs.rs
@@ -1,4 +1,4 @@
-use backend::{bail_span, BindgenResult, Diagnostic};
+use napi_macro_backend::{bail_span, BindgenResult, Diagnostic};
 use proc_macro2::{Delimiter, Ident, Span, TokenTree};
 use std::{
   cell::{Cell, RefCell},
