+++ examples-compat-mode/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+authors = ["LongYinan <lynweklm@gmail.com>"]
+edition = "2018"
+name = "napi-examples-with-compat-mode"
+version = "0.1.0"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+napi-derive = { path = "../crates/macro", features = ["compat-mode"] }
+napi = { path = "../napi", features = ["compat-mode"] }
+
+[build-dependencies]
+napi-build = { path = "../build" }
