+++ examples-compat-mode/build.rs
