+++ examples/build.rs
