+++ examples/src/lib.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+extern crate napi_derive;
+
+mod number;
+mod string;
