+++ examples/src/string.rs
@@ -1,0 +1,6 @@
+use napi::bindgen_prelude::*;
+
+#[napi]
+fn contains(source: String, target: String) -> bool {
+  source.contains(&target)
+}
