+++ memory-testing/Cargo.toml
@@ -9,8 +9,8 @@ crate-type = ["cdylib"]
 
 [dependencies]
 futures = "0.3"
-napi = {path = "../napi", features = ["tokio_rt", "serde-json", "latin1"]}
-napi-derive = {path = "../napi-derive"}
+napi = {path = "../napi", features = ["tokio_rt", "serde-json", "latin1", "compat-mode"]}
+napi-derive = {path = "../crates/macro", features = ["compat-mode"]}
 serde = "1"
 serde_bytes = "0.11"
 serde_derive = "1"
