--- napi-derive/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-authors = ["LongYinan <lynweklm@gmail.com>"]
-description = "N-API procedural macros"
-edition = "2018"
-keywords = ["NodeJS", "FFI", "NAPI", "n-api"]
-license = "MIT"
-name = "napi-derive"
-readme = "README.md"
-repository = "https://github.com/napi-rs/napi-rs"
-version = "1.1.0"
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = {version = "1.0", features = ["fold", "full"]}
-
-[lib]
-proc-macro = true
