+++ napi/src/value_type.rs
