+++ test_module/Cargo.toml
@@ -13,8 +13,8 @@ napi3 = ["napi/napi3"]
 
 [dependencies]
 futures = "0.3"
-napi = {path = "../napi", features = ["tokio_rt", "serde-json", "latin1"]}
-napi-derive = {path = "../napi-derive"}
+napi = {path = "../napi", features = ["tokio_rt", "serde-json", "latin1", "compat-mode"]}
+napi-derive = {path = "../crates/macro", features = ["compat-mode"]}
 serde = "1"
 serde_bytes = "0.11"
 serde_derive = "1"
